+++ src/main/resources/eidos.conf
@@ -22,7 +22,7 @@ EidosSystem {
     geoWord2IdxPath = /org/clulab/wm/eidos/english/context/word2idx_file.txt
       geoLoc2IdPath = /org/clulab/wm/eidos/english/context/geo_dict_with_population_SOUTH_SUDAN.txt
 
-  cacheDir = ./cache/${EidosSystem.language}
+           cacheDir = ./cache/${EidosSystem.language}
              useW2V = false
         useTimeNorm = false
          useGeoNorm = true
