+++ src/main/scala/org/clulab/wm/eidos/context/GeoPhraseID.scala
@@ -1,0 +1,5 @@
+package org.clulab.wm.eidos.context
+
+case class GeoPhraseID(phraseID: String, PhraseGeoID: Option[Int], StartOffset_locs: Int, EndOffset_locs: Int)
+
+
